Add files via upload
It is a menu-driven program that allows customers to make purchases and Admins to update and view data. The project displays various mobile phones with details such as price, model, and release year. Prices can be sorted in ascending or descending order to aid customer selection. The program utilizes linked lists, binary trees, binary search trees, sorting, and searching algorithms.